camera/requirements.txt
# Python 3.10–3.12 recommended (Windows-friendly)
opencv-python>=4.8.0,<5
mediapipe>=0.10.0
numpy>=1.24.0,<3
pyserial>=3.5
open-rppg>=0.1.0
